+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [2.2.0] - 2021-07-06
+
+### Changed
+
+- `HttpClient` logs additional request data (query parameters, body).
+
+### Added
+
+- `HttpClientOptions` now accepts `LogOptions` object that allows enabling informational request and **response (new)** logs.
+
+```js
+{
+  logOptions: {
+    enabledLogs: [LogType.requests, LogType.responses];
+  }
+}
+```
+
 ## [2.1.0] - 2021-03-11
 
 ### Changed